+++ src/containers/CouponDetail.tsx
@@ -10,11 +10,14 @@ import Moment from 'react-moment';
 import axios from 'axios';
 let fileDownload = require('js-file-download');
 const pdfcontract = require('../assets/Token-Legal-Contract_Placeholder.pdf');
+const isDev = process.env.NODE_ENV === 'development';
 
 const CouponDetail: React.FC = () => {
   const { t } = useTranslation();
   const history = useHistory();
   const [isMintable, setIsMintable] = useState(false);
+  const [isDownloading, setIsDownloading] = useState(false);
+  const [error, setError] = useState('');
 
   const { ethAddress, token } = useSelector(({ wallet }: any) => {
     return {
@@ -39,11 +42,16 @@ const CouponDetail: React.FC = () => {
   }, [ethAddress, history, isMintable, token]);
 
   const handleDownload = (url: string, filename: string) => {
-    axios.get(url, {
+    setIsDownloading(true);
+    axios.get(isDev ? url : url.replace('http',  'https'), {
       responseType: 'blob',
     })
     .then((res) => {
+      setIsDownloading(false);
       fileDownload(res.data, filename)
+    }).catch(() => {
+      setIsDownloading(false);
+      setError(t('common.download_error'))
     })
   }
 
@@ -141,7 +149,7 @@ const CouponDetail: React.FC = () => {
             with a total supply of {token?.amount} coupons
           </Text>
         </Flex>
-        <Flex paddingX="15px" paddingBottom={10}>
+        <Flex paddingX="15px" paddingBottom={15} marginBottom={(isDownloading || error) ? '80px' : '30px' }>
           <button onClick={() => handleDownload(token?.contractHash || pdfcontract, token?.contractLabel || 'Token-Legal-Contract_Placeholder.pdf')}>
             <Flex
               marginTop="10px"
@@ -158,6 +166,12 @@ const CouponDetail: React.FC = () => {
             </Flex>
           </button>
         </Flex>
+        {isDownloading && <Flex padding='10px 20px' margin='auto' justifyContent="center" width='150px' sx={{ background: 'rgb(47 47 47 / 77%)', borderRadius: '25px', position: 'absolute', bottom: '30px', left: 0, right: 0 }}>
+          <Text color="#ffffff">{t('common.downloading')}</Text>
+        </Flex>}
+        {error && <Flex padding='10px 20px' margin='auto' justifyContent="center" width='150px' sx={{ background: '#DD303D', borderRadius: '25px', position: 'absolute', bottom: '30px', left: 0, right: 0 }}>
+          <Text color="#ffffff">{error}</Text>
+        </Flex>}
       </Flex>
     </Flex>
   );